refactor(audit): split shell and source layers into sibling modules
audit.rs had grown to ~4.2k lines with two self-contained seams
inside it. Move the shell tokenizer/target-extraction layer to
audit_shell.rs and the action source selection-and-fetch layer to
audit_source.rs, each taking the unit tests that exercise it
directly (6 and 23 tests respectively; the 549-test total is
unchanged).

Pure code motion: no logic, signature, or output changes beyond
pub(crate) visibility on the seam functions and the new module
docs. The public audit surface (run, extract_run_blocks,
scan_shell_content) stays in audit.rs, so score.rs is untouched;
main.rs only gains the module declarations.
