+++ src/audit.rs
@@ -1,10 +1,8 @@
 use anyhow::{Context, Result};
 use serde_norway::Value;
 use std::collections::HashSet;
-use std::path::{Component, Path, PathBuf};
+use std::path::Path;
 use std::process::ExitCode;
-use std::sync::Arc;
-use tokio::sync::Semaphore;
 
 use crate::audit_patterns::{
     self, DOCKER_PATTERNS, DOCKER_URL_PATTERNS, JS_PATTERNS, JS_URL_PATTERNS,
@@ -13,15 +11,23 @@ use crate::audit_patterns::{
     SH_NPM_UNVERSIONED, SH_NPX_UNVERSIONED, SH_PIP_GIT_URL_UNVERSIONED, SH_PIP_UNVERSIONED,
     SHELL_PATTERNS, SHELL_PIPE_PATTERNS, SHELL_URL_PATTERNS, cargo_install_has_version,
     category_str, extract_urls, gem_install_has_version, gh_release_has_tag,
-    git_clone_has_pinned_ref, has_checksum_verify, npm_install_has_version, npx_has_version,
-    pip_git_url_has_ref, pip_install_has_version, ps_install_has_required_version, url_has_version,
+    git_clone_has_pinned_ref, npm_install_has_version, npx_has_version, pip_git_url_has_ref,
+    pip_install_has_version, ps_install_has_required_version, url_has_version,
+};
+use crate::audit_shell::{
+    checksum_verifies_target, fetch_output_targets, git_clone_has_bound_sha_checkout,
+    is_shell_comment_line, join_continuations, url_piped_to_jq,
+};
+use crate::audit_source::{
+    ActionScanStatus, remote_action_scan_key, scan_action_source, scan_local_action_source,
+    short_sha,
 };
 use crate::audited_actions::{AuditSource, AuditedActions};
 use crate::auth;
 use crate::config::Config;
 use crate::github::GitHubClient;
 use crate::output::{self, AuditFinding, AuditMatch, AuditReport};
-use crate::workflow::{self, ActionRef, LocalActionRef};
+use crate::workflow;
 use colored::Colorize;
 
 /// Reason string for matches allowed via the `trusted-hosts` config list.
@@ -363,10 +369,6 @@ pub async fn run(
     }
 }
 
-fn remote_action_scan_key(action: &ActionRef) -> String {
-    format!("{}@{}", action.full_name(), action.ref_string)
-}
-
 pub fn extract_run_blocks(path: &Path, content: &str) -> Result<Vec<(usize, String)>> {
     let yaml: Value =
         serde_norway::from_str(content).with_context(|| format!("parsing {}", path.display()))?;
@@ -403,7 +405,7 @@ pub fn extract_run_blocks(path: &Path, content: &str) -> Result<Vec<(usize, Stri
 /// `parallel:` nested inside `parallel:`. A `background: true` step keeps its
 /// own `run:` key and is collected like any other step. Steps without a `run:`
 /// (`uses:`, `wait:`, `cancel:`) contribute nothing.
-fn collect_step_run_blocks(steps: &Value) -> Vec<&str> {
+pub(crate) fn collect_step_run_blocks(steps: &Value) -> Vec<&str> {
     fn collect<'a>(steps: &'a Value, runs: &mut Vec<&'a str>) {
         let Some(sequence) = steps.as_sequence() else {
             return;
@@ -461,588 +463,6 @@ fn find_run_line(file_content: &str, run_content: &str, start: usize) -> (usize,
         Some(i) => (i + 1, i + 1),
         None => (0, start),
     }
-}
-
-/// Whether a shell source line is a pure comment and thus never executed.
-/// Trailing comments on a command line are not covered — stripping an
-/// unquoted `#` would require full shell tokenization and risks hiding a
-/// real payload embedded in a quoted string.
-fn is_shell_comment_line(line: &str) -> bool {
-    line.trim_start().starts_with('#')
-}
-
-/// Join shell lines ending in `\` into a single logical line, anchored at the
-/// 0-based index of the first physical line.
-fn join_continuations(content: &str) -> Vec<(usize, String)> {
-    let mut out: Vec<(usize, String)> = Vec::new();
-    let mut pending: Option<(usize, String)> = None;
-    for (i, raw) in content.lines().enumerate() {
-        let trimmed_end = raw.trim_end();
-        // A comment never continues onto the next line — a trailing backslash
-        // is just comment text, so the next line is a new command. Joining it
-        // into the comment would hide it from the scan.
-        let backslash_continuation = trimmed_end.ends_with('\\');
-        let operator_continuation = ends_with_pipeline_operator(trimmed_end);
-        let is_continuation = (backslash_continuation || operator_continuation)
-            && !(pending.is_none() && is_shell_comment_line(raw));
-        let body = if backslash_continuation {
-            &trimmed_end[..trimmed_end.len() - 1]
-        } else {
-            raw
-        };
-        match pending.as_mut() {
-            Some((_, buf)) => {
-                buf.push(' ');
-                buf.push_str(body.trim());
-            }
-            None => {
-                pending = Some((i, body.to_string()));
-            }
-        }
-        if !is_continuation && let Some((start, buf)) = pending.take() {
-            out.push((start, buf));
-        }
-    }
-    if let Some(p) = pending {
-        out.push(p);
-    }
-    out
-}
-
-fn ends_with_pipeline_operator(line: &str) -> bool {
-    line.ends_with('|') || line.ends_with("&&") || line.ends_with("||")
-}
-
-#[derive(Debug, Clone)]
-struct ShellCommand {
-    text: String,
-    stages: Vec<ShellStage>,
-}
-
-#[derive(Debug, Clone)]
-struct ShellStage {
-    text: String,
-    words: Vec<String>,
-}
-
-#[derive(Debug, Clone)]
-struct GitCloneCommand {
-    dir: String,
-    command_index: usize,
-}
-
-fn parse_shell_line(line: &str) -> Vec<ShellCommand> {
-    split_shell_control(line)
-        .into_iter()
-        .filter_map(|command| {
-            let stages: Vec<ShellStage> = split_shell_pipeline(&command)
-                .into_iter()
-                .map(|stage| ShellStage {
-                    words: shell_words(&stage),
-                    text: stage,
-                })
-                .collect();
-            if stages.is_empty() {
-                None
-            } else {
-                Some(ShellCommand {
-                    text: command,
-                    stages,
-                })
-            }
-        })
-        .collect()
-}
-
-fn split_shell_control(line: &str) -> Vec<String> {
-    split_shell(line, SplitMode::Control)
-}
-
-fn split_shell_pipeline(line: &str) -> Vec<String> {
-    split_shell(line, SplitMode::Pipeline)
-}
-
-enum SplitMode {
-    Control,
-    Pipeline,
-}
-
-fn split_shell(line: &str, mode: SplitMode) -> Vec<String> {
-    let mut out = Vec::new();
-    let mut current = String::new();
-    let mut chars = line.chars().peekable();
-    let mut quote: Option<char> = None;
-    let mut escaped = false;
-
-    while let Some(ch) = chars.next() {
-        if escaped {
-            current.push(ch);
-            escaped = false;
-            continue;
-        }
-        if ch == '\\' && quote != Some('\'') {
-            current.push(ch);
-            escaped = true;
-            continue;
-        }
-        if let Some(q) = quote {
-            current.push(ch);
-            if ch == q {
-                quote = None;
-            }
-            continue;
-        }
-        if ch == '\'' || ch == '"' {
-            quote = Some(ch);
-            current.push(ch);
-            continue;
-        }
-
-        match mode {
-            SplitMode::Control if ch == ';' => {
-                push_shell_part(&mut out, &mut current);
-            }
-            SplitMode::Control if ch == '&' && chars.peek() == Some(&'&') => {
-                chars.next();
-                push_shell_part(&mut out, &mut current);
-            }
-            SplitMode::Control if ch == '|' && chars.peek() == Some(&'|') => {
-                chars.next();
-                push_shell_part(&mut out, &mut current);
-            }
-            SplitMode::Pipeline if ch == '|' && chars.peek() != Some(&'|') => {
-                push_shell_part(&mut out, &mut current);
-            }
-            _ => current.push(ch),
-        }
-    }
-
-    push_shell_part(&mut out, &mut current);
-    out
-}
-
-fn push_shell_part(out: &mut Vec<String>, current: &mut String) {
-    let part = current.trim();
-    if !part.is_empty() {
-        out.push(part.to_string());
-    }
-    current.clear();
-}
-
-fn shell_words(command: &str) -> Vec<String> {
-    let mut words = Vec::new();
-    let mut current = String::new();
-    let mut chars = command.chars().peekable();
-    let mut quote: Option<char> = None;
-    let mut escaped = false;
-
-    while let Some(ch) = chars.next() {
-        if escaped {
-            current.push(ch);
-            escaped = false;
-            continue;
-        }
-        if ch == '\\' && quote != Some('\'') {
-            escaped = true;
-            continue;
-        }
-        if let Some(q) = quote {
-            if ch == q {
-                quote = None;
-            } else {
-                current.push(ch);
-            }
-            continue;
-        }
-        if ch == '\'' || ch == '"' {
-            quote = Some(ch);
-            continue;
-        }
-        if ch.is_whitespace() {
-            push_shell_word(&mut words, &mut current);
-            continue;
-        }
-        if ch == '>' {
-            push_shell_word(&mut words, &mut current);
-            if chars.peek() == Some(&'>') {
-                chars.next();
-                words.push(">>".to_string());
-            } else {
-                words.push(">".to_string());
-            }
-            continue;
-        }
-        current.push(ch);
-    }
-
-    push_shell_word(&mut words, &mut current);
-    words
-}
-
-fn push_shell_word(words: &mut Vec<String>, current: &mut String) {
-    if !current.is_empty() {
-        words.push(std::mem::take(current));
-    }
-}
-
-fn url_piped_to_jq(line: &str, url: &str) -> bool {
-    parse_shell_line(line).into_iter().any(|command| {
-        let Some(fetch_stage) = command
-            .stages
-            .iter()
-            .position(|stage| stage.text.contains(url))
-        else {
-            return false;
-        };
-        command
-            .stages
-            .iter()
-            .skip(fetch_stage + 1)
-            .any(stage_invokes_jq)
-    })
-}
-
-fn stage_invokes_jq(stage: &ShellStage) -> bool {
-    command_word_index(&stage.words).is_some_and(|idx| stage.words[idx] == "jq")
-}
-
-fn command_word_index(words: &[String]) -> Option<usize> {
-    words
-        .iter()
-        .position(|word| !word.contains('=') || word.starts_with('-'))
-}
-
-fn fetch_output_targets(line: &str) -> Vec<String> {
-    let mut targets = Vec::new();
-    for command in parse_shell_line(line) {
-        for stage in command.stages {
-            if let Some(target) = fetch_output_target(&stage)
-                && !targets.contains(&target)
-            {
-                targets.push(target);
-            }
-        }
-    }
-    targets
-}
-
-fn fetch_output_target(stage: &ShellStage) -> Option<String> {
-    let fetch_index = stage
-        .words
-        .iter()
-        .position(|word| word == "curl" || word == "wget")?;
-    let fetch = stage.words[fetch_index].as_str();
-    let option_target = match fetch {
-        "curl" => curl_output_target(&stage.words[fetch_index + 1..]),
-        "wget" => wget_output_target(&stage.words[fetch_index + 1..]),
-        _ => None,
-    };
-    option_target.or_else(|| redirect_output_target(&stage.words))
-}
-
-fn curl_output_target(words: &[String]) -> Option<String> {
-    let mut urls = words
-        .iter()
-        .filter(|word| word.starts_with("http://") || word.starts_with("https://"));
-    let mut i = 0;
-    while i < words.len() {
-        let word = words[i].as_str();
-        if matches!(word, "-o" | "--output") {
-            return words.get(i + 1).and_then(|target| usable_target(target));
-        }
-        if let Some(target) = word.strip_prefix("--output=") {
-            return usable_target(target);
-        }
-        if word == "-O" || word == "--remote-name" || short_flag_has_remote_name(word) {
-            return urls.next().and_then(|url| url_basename(url));
-        }
-        if short_flag_uses_output(word) {
-            return words.get(i + 1).and_then(|target| usable_target(target));
-        }
-        if let Some(target) = curl_attached_output(word) {
-            return usable_target(target);
-        }
-        i += 1;
-    }
-    None
-}
-
-fn curl_attached_output(word: &str) -> Option<&str> {
-    let flags = word.strip_prefix('-')?;
-    if flags.starts_with('-') {
-        return None;
-    }
-    let pos = flags.find('o')?;
-    let target = &flags[pos + 1..];
-    (!target.is_empty()).then_some(target)
-}
-
-fn short_flag_uses_output(word: &str) -> bool {
-    let Some(flags) = word.strip_prefix('-') else {
-        return false;
-    };
-    !flags.starts_with('-') && flags.ends_with('o')
-}
-
-fn short_flag_has_remote_name(word: &str) -> bool {
-    let Some(flags) = word.strip_prefix('-') else {
-        return false;
-    };
-    !flags.starts_with('-') && flags.contains('O')
-}
-
-fn wget_output_target(words: &[String]) -> Option<String> {
-    let mut i = 0;
-    while i < words.len() {
-        let word = words[i].as_str();
-        if matches!(word, "-O" | "--output-document") {
-            return words.get(i + 1).and_then(|target| usable_target(target));
-        }
-        if let Some(target) = word.strip_prefix("--output-document=") {
-            return usable_target(target);
-        }
-        if let Some(target) = word.strip_prefix("-O")
-            && !target.is_empty()
-        {
-            return usable_target(target);
-        }
-        i += 1;
-    }
-    None
-}
-
-fn redirect_output_target(words: &[String]) -> Option<String> {
-    words.windows(2).find_map(|pair| {
-        if pair[0] == ">" || pair[0] == ">>" {
-            usable_target(&pair[1])
-        } else {
-            None
-        }
-    })
-}
-
-fn usable_target(target: &str) -> Option<String> {
-    let target = normalize_path_token(target);
-    (!target.is_empty() && target != "-" && !target.starts_with('&')).then_some(target)
-}
-
-fn url_basename(url: &str) -> Option<String> {
-    let clean = url.split(['?', '#']).next().unwrap_or(url);
-    let name = clean.rsplit('/').next().unwrap_or_default();
-    usable_target(name)
-}
-
-fn checksum_verifies_target(line: &str, target: &str) -> bool {
-    parse_shell_line(line).into_iter().any(|command| {
-        command
-            .stages
-            .iter()
-            .any(|stage| checksum_stage_verifies_target(stage, target))
-    })
-}
-
-fn checksum_stage_verifies_target(stage: &ShellStage, target: &str) -> bool {
-    has_checksum_verify(&stage.text)
-        && stage
-            .words
-            .iter()
-            .any(|word| checksum_word_matches_target(word, target))
-}
-
-fn checksum_word_matches_target(word: &str, target: &str) -> bool {
-    let word = normalize_path_token(word);
-    let target = normalize_path_token(target);
-    word == target
-        || word
-            .strip_prefix(&target)
-            .is_some_and(|suffix| matches!(suffix, ".sha256" | ".sha512" | ".sha1" | ".sig"))
-}
-
-fn git_clone_has_bound_sha_checkout(logical: &[(usize, String)], li: usize) -> bool {
-    let clones = unpinned_git_clones(&logical[li].1);
-    !clones.is_empty()
-        && clones
-            .iter()
-            .all(|clone| clone_has_bound_sha_checkout(logical, li, clone))
-}
-
-fn unpinned_git_clones(line: &str) -> Vec<GitCloneCommand> {
-    parse_shell_line(line)
-        .into_iter()
-        .enumerate()
-        .filter_map(|(command_index, command)| {
-            if command.stages.iter().any(stage_has_git_clone)
-                && !git_clone_has_pinned_ref(&command.text)
-            {
-                git_clone_dir(&command).map(|dir| GitCloneCommand { dir, command_index })
-            } else {
-                None
-            }
-        })
-        .collect()
-}
-
-fn stage_has_git_clone(stage: &ShellStage) -> bool {
-    git_word_index(&stage.words)
-        .is_some_and(|idx| stage.words.get(idx + 1).is_some_and(|w| w == "clone"))
-}
-
-fn git_clone_dir(command: &ShellCommand) -> Option<String> {
-    let stage = command
-        .stages
-        .iter()
-        .find(|stage| stage_has_git_clone(stage))?;
-    let git = git_word_index(&stage.words)?;
-    let args = &stage.words[git + 2..];
-    let mut positionals = Vec::new();
-    let mut i = 0;
-    while i < args.len() {
-        let word = args[i].as_str();
-        if git_clone_option_takes_value(word) {
-            i += 2;
-            continue;
-        }
-        if let Some(flag) = word.split_once('=').map(|(flag, _)| flag)
-            && git_clone_option_takes_value(flag)
-        {
-            i += 1;
-            continue;
-        }
-        if word.starts_with('-') {
-            i += 1;
-            continue;
-        }
-        positionals.push(word);
-        i += 1;
-    }
-
-    if let Some(dir) = positionals.get(1) {
-        usable_target(dir)
-    } else {
-        positionals.first().and_then(|url| clone_default_dir(url))
-    }
-}
-
-fn git_clone_option_takes_value(flag: &str) -> bool {
-    matches!(
-        flag,
-        "-b" | "--branch"
-            | "-c"
-            | "--config"
-            | "--depth"
-            | "--origin"
-            | "-o"
-            | "--template"
-            | "--reference"
-            | "--reference-if-able"
-            | "--separate-git-dir"
-            | "--jobs"
-            | "-j"
-    )
-}
-
-fn clone_default_dir(url: &str) -> Option<String> {
-    let clean = url.trim_end_matches('/').trim_end_matches(".git");
-    usable_target(clean.rsplit(['/', ':']).next().unwrap_or(clean))
-}
-
-fn clone_has_bound_sha_checkout(
-    logical: &[(usize, String)],
-    clone_line: usize,
-    clone: &GitCloneCommand,
-) -> bool {
-    let mut current_dir: Option<String> = None;
-    for offset in 0..=3 {
-        let Some((_, line)) = logical.get(clone_line + offset) else {
-            break;
-        };
-        let commands = parse_shell_line(line);
-        for (command_index, command) in commands.iter().enumerate() {
-            if offset == 0 && command_index <= clone.command_index {
-                continue;
-            }
-            if let Some(dir) = command_cd_dir(command) {
-                current_dir = Some(dir);
-                continue;
-            }
-            if command_has_bound_checkout(command, &clone.dir, current_dir.as_deref()) {
-                return true;
-            }
-        }
-    }
-    false
-}
-
-fn command_cd_dir(command: &ShellCommand) -> Option<String> {
-    let stage = command.stages.first()?;
-    let idx = command_word_index(&stage.words)?;
-    if stage.words[idx] == "cd" {
-        stage.words.get(idx + 1).and_then(|dir| usable_target(dir))
-    } else {
-        None
-    }
-}
-
-fn command_has_bound_checkout(
-    command: &ShellCommand,
-    clone_dir: &str,
-    current_dir: Option<&str>,
-) -> bool {
-    command.stages.iter().any(|stage| {
-        git_checkout_sha_dir(stage).is_some_and(|checkout_dir| match checkout_dir {
-            Some(dir) => same_shell_path(&dir, clone_dir),
-            None => current_dir.is_some_and(|dir| same_shell_path(dir, clone_dir)),
-        })
-    })
-}
-
-fn git_checkout_sha_dir(stage: &ShellStage) -> Option<Option<String>> {
-    let git = git_word_index(&stage.words)?;
-    let mut checkout_dir = None;
-    let mut i = git + 1;
-    while i < stage.words.len() {
-        let word = stage.words[i].as_str();
-        if word == "-C" {
-            checkout_dir = stage.words.get(i + 1).and_then(|dir| usable_target(dir));
-            i += 2;
-            continue;
-        }
-        if let Some(dir) = word.strip_prefix("-C")
-            && !dir.is_empty()
-        {
-            checkout_dir = usable_target(dir);
-            i += 1;
-            continue;
-        }
-        if word == "checkout" && stage.words.get(i + 1).is_some_and(|sha| is_full_sha(sha)) {
-            return Some(checkout_dir);
-        }
-        i += 1;
-    }
-    None
-}
-
-fn git_word_index(words: &[String]) -> Option<usize> {
-    words.iter().position(|word| word == "git")
-}
-
-fn is_full_sha(word: &str) -> bool {
-    word.len() == 40 && word.chars().all(|c| c.is_ascii_hexdigit())
-}
-
-fn same_shell_path(left: &str, right: &str) -> bool {
-    normalize_path_token(left) == normalize_path_token(right)
-}
-
-fn normalize_path_token(path: &str) -> String {
-    let mut path = path
-        .trim_matches(|c| matches!(c, '"' | '\'' | ')' | '(' | ',' | ';'))
-        .trim()
-        .to_string();
-    while let Some(rest) = path.strip_prefix("./") {
-        path = rest.to_string();
-    }
-    path
 }
 
 pub fn scan_shell_content(
@@ -1283,7 +703,7 @@ pub fn scan_shell_content(
 /// Lines longer than this are treated as minified and split on `;` before scanning.
 const MINIFIED_LINE_THRESHOLD: usize = 500;
 
-fn scan_js_content(
+pub(crate) fn scan_js_content(
     content: &str,
     source_file: &str,
     action_name: &str,
@@ -1339,7 +759,7 @@ fn scan_js_content(
     }
 }
 
-fn scan_py_content(
+pub(crate) fn scan_py_content(
     content: &str,
     source_file: &str,
     action_name: &str,
@@ -1369,7 +789,7 @@ fn scan_py_content(
     }
 }
 
-fn scan_dockerfile_content(
+pub(crate) fn scan_dockerfile_content(
     content: &str,
     source_file: &str,
     action_name: &str,
@@ -1581,481 +1001,6 @@ fn check_patterns(
             });
         }
     }
-}
-
-/// Third-party dependency dirs, not the action's own source. An action that
-/// commits `node_modules/` would otherwise cost thousands of per-file fetches
-/// for zero signal. `dist/` is excluded — a bundled `dist/index.js` is the
-/// code that actually runs and must still be scanned.
-const VENDORED_DIRS: &[&str] = &["node_modules", "site-packages", ".venv", "venv"];
-
-/// Matches whole path components, so `node_modules_helper.js` is not affected.
-fn is_vendored_path(path: &str) -> bool {
-    path.split('/')
-        .any(|component| VENDORED_DIRS.contains(&component))
-}
-
-/// Max action source files fetched concurrently. Bounds the fan-out so a
-/// file-heavy action doesn't burst into a rate-limit-exhausting wave.
-const MAX_CONCURRENT_FILE_FETCHES: usize = 8;
-
-/// Which scanner handles a given action source file.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum SourceFileKind {
-    ActionYml,
-    JavaScript,
-    Python,
-    Dockerfile,
-}
-
-/// Whether all selected action source files were fetched and parsed well
-/// enough to support a durable "clean" verdict.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum ActionScanStatus {
-    Complete,
-    Incomplete,
-}
-
-impl ActionScanStatus {
-    fn from_complete(complete: bool) -> Self {
-        if complete {
-            Self::Complete
-        } else {
-            Self::Incomplete
-        }
-    }
-}
-
-/// Select the files in a fetched action tree worth scanning, each paired with
-/// its scanner, in tree order. Pure (no I/O) so the filtering is unit testable.
-fn select_source_files(
-    tree: &[crate::github::TreeEntry],
-    base: &str,
-) -> Vec<(String, SourceFileKind)> {
-    let mut targets = Vec::new();
-    for entry in tree {
-        if entry.entry_type != "blob" {
-            continue;
-        }
-        let path = &entry.path;
-        if is_vendored_path(path) {
-            continue;
-        }
-        if !base.is_empty() && path != base && !path.starts_with(&format!("{base}/")) {
-            continue;
-        }
-        let relative = if base.is_empty() {
-            path.as_str()
-        } else {
-            path.strip_prefix(base)
-                .unwrap_or(path)
-                .trim_start_matches('/')
-        };
-        let kind = if relative == "action.yml" || relative == "action.yaml" {
-            SourceFileKind::ActionYml
-        } else if is_javascript_source(path) {
-            SourceFileKind::JavaScript
-        } else if path.ends_with(".py") {
-            SourceFileKind::Python
-        } else if relative == "Dockerfile" || path.ends_with(".dockerfile") {
-            SourceFileKind::Dockerfile
-        } else {
-            continue;
-        };
-        targets.push((path.clone(), kind));
-    }
-    targets
-}
-
-fn force_include_remote_action_entrypoints(
-    base: &str,
-    targets: &mut Vec<(String, SourceFileKind)>,
-    contents: &[Option<Result<String>>],
-) {
-    let metadata: Vec<String> = targets
-        .iter()
-        .zip(contents)
-        .filter_map(|((_, kind), content)| {
-            if *kind != SourceFileKind::ActionYml {
-                return None;
-            }
-            let Some(Ok(content)) = content.as_ref() else {
-                return None;
-            };
-            Some(content.clone())
-        })
-        .collect();
-
-    for content in metadata {
-        let Ok(yaml) = serde_norway::from_str::<Value>(&content) else {
-            continue;
-        };
-        for entrypoint in action_yml_entrypoint_paths(&yaml, base) {
-            push_unique_source_target(targets, entrypoint, SourceFileKind::JavaScript);
-        }
-    }
-}
-
-fn force_include_local_action_entrypoints(
-    action_dir: &Path,
-    targets: &mut Vec<(PathBuf, SourceFileKind)>,
-) {
-    let metadata_paths: Vec<PathBuf> = targets
-        .iter()
-        .filter_map(|(path, kind)| (*kind == SourceFileKind::ActionYml).then_some(path.clone()))
-        .collect();
-
-    for path in metadata_paths {
-        let Ok(content) = std::fs::read_to_string(&path) else {
-            continue;
-        };
-        let Ok(yaml) = serde_norway::from_str::<Value>(&content) else {
-            continue;
-        };
-        for entrypoint in action_yml_entrypoint_paths(&yaml, "") {
-            let path = action_dir.join(&entrypoint);
-            if local_entrypoint_is_regular_file(&path) {
-                push_unique_local_source_target(targets, path, SourceFileKind::JavaScript);
-            }
-        }
-    }
-}
-
-fn action_yml_entrypoint_paths(yaml: &Value, base: &str) -> Vec<String> {
-    let Some(runs) = yaml.get("runs").and_then(|r| r.as_mapping()) else {
-        return Vec::new();
-    };
-
-    ["main", "pre", "post"]
-        .into_iter()
-        .filter_map(|key| runs.get(key).and_then(|v| v.as_str()))
-        .filter_map(|path| normalize_action_entrypoint_path(base, path))
-        .collect()
-}
-
-fn normalize_action_entrypoint_path(base: &str, path: &str) -> Option<String> {
-    let path = path.trim();
-    if path.is_empty() || path.contains('\\') {
-        return None;
-    }
-    let rel = Path::new(path);
-    if rel.is_absolute() {
-        return None;
-    }
-
-    let mut parts = Vec::new();
-    for component in rel.components() {
-        match component {
-            Component::Normal(part) => parts.push(part.to_str()?.to_string()),
-            Component::CurDir => {}
-            _ => return None,
-        }
-    }
-    if parts.is_empty() {
-        return None;
-    }
-
-    let rel = parts.join("/");
-    if base.is_empty() {
-        Some(rel)
-    } else {
-        Some(format!("{base}/{rel}"))
-    }
-}
-
-fn push_unique_source_target(
-    targets: &mut Vec<(String, SourceFileKind)>,
-    path: String,
-    kind: SourceFileKind,
-) {
-    if !targets.iter().any(|(existing, _)| existing == &path) {
-        targets.push((path, kind));
-    }
-}
-
-fn push_unique_local_source_target(
-    targets: &mut Vec<(PathBuf, SourceFileKind)>,
-    path: PathBuf,
-    kind: SourceFileKind,
-) {
-    if !targets.iter().any(|(existing, _)| existing == &path) {
-        targets.push((path, kind));
-    }
-}
-
-fn local_entrypoint_is_regular_file(path: &Path) -> bool {
-    std::fs::symlink_metadata(path)
-        .map(|metadata| metadata.file_type().is_file())
-        .unwrap_or_default()
-}
-
-async fn fetch_remote_source_files(
-    client: &GitHubClient,
-    action: &ActionRef,
-    targets: &[(String, SourceFileKind)],
-) -> (Vec<Option<Result<String>>>, bool) {
-    let semaphore = Arc::new(Semaphore::new(MAX_CONCURRENT_FILE_FETCHES));
-    let mut fetches = tokio::task::JoinSet::new();
-    for (index, (path, _)) in targets.iter().enumerate() {
-        let client = client.clone();
-        let owner = action.owner.clone();
-        let repo = action.repo.clone();
-        let git_ref = action.ref_string.clone();
-        let path = path.clone();
-        let semaphore = Arc::clone(&semaphore);
-        fetches.spawn(async move {
-            let _permit = semaphore.acquire_owned().await.ok();
-            let content = client.fetch_file(&owner, &repo, &path, &git_ref).await;
-            (index, content)
-        });
-    }
-
-    let mut contents: Vec<Option<Result<String>>> = (0..targets.len()).map(|_| None).collect();
-    let mut complete = true;
-    while let Some(joined) = fetches.join_next().await {
-        match joined {
-            Ok((index, content)) => {
-                if content.is_err() {
-                    complete = false;
-                }
-                contents[index] = Some(content);
-            }
-            Err(_) => {
-                complete = false;
-            }
-        }
-    }
-
-    (contents, complete)
-}
-
-fn collect_local_source_files(action_dir: &Path) -> Result<Vec<(PathBuf, SourceFileKind)>> {
-    fn visit(dir: &Path, base: &Path, targets: &mut Vec<(PathBuf, SourceFileKind)>) -> Result<()> {
-        let mut entries = Vec::new();
-        for entry in std::fs::read_dir(dir).with_context(|| format!("reading {}", dir.display()))? {
-            entries.push(entry?);
-        }
-        entries.sort_by_key(|e| e.path());
-
-        for entry in entries {
-            let path = entry.path();
-            let relative = path
-                .strip_prefix(base)
-                .unwrap_or(&path)
-                .to_string_lossy()
-                .replace('\\', "/");
-
-            if is_vendored_path(&relative) {
-                continue;
-            }
-
-            // `file_type()` does not follow symlinks, so a symlinked entry is
-            // neither file nor dir here and is skipped. Deliberate: it keeps
-            // traversal inside the action directory (a symlink can't redirect
-            // the scan outside it) at the cost of not scanning symlinked source
-            // — an acceptable trade-off since local actions are first-party.
-            let file_type = entry.file_type()?;
-            if file_type.is_dir() {
-                visit(&path, base, targets)?;
-            } else if file_type.is_file()
-                && let Some(kind) = source_file_kind(&relative)
-            {
-                targets.push((path, kind));
-            }
-        }
-        Ok(())
-    }
-
-    let mut targets = Vec::new();
-    visit(action_dir, action_dir, &mut targets)?;
-    Ok(targets)
-}
-
-fn source_file_kind(relative: &str) -> Option<SourceFileKind> {
-    if relative == "action.yml" || relative == "action.yaml" {
-        Some(SourceFileKind::ActionYml)
-    } else if is_javascript_source(relative) {
-        Some(SourceFileKind::JavaScript)
-    } else if relative.ends_with(".py") {
-        Some(SourceFileKind::Python)
-    } else if relative == "Dockerfile" || relative.ends_with(".dockerfile") {
-        Some(SourceFileKind::Dockerfile)
-    } else {
-        None
-    }
-}
-
-fn is_javascript_source(path: &str) -> bool {
-    path.ends_with(".js")
-        || path.ends_with(".ts")
-        || path.ends_with(".mjs")
-        || path.ends_with(".cjs")
-}
-
-fn local_action_dir(repo_root: &Path, action: &LocalActionRef) -> Result<PathBuf> {
-    let Some(rel) = action.path.strip_prefix("./") else {
-        anyhow::bail!("local action path must start with ./");
-    };
-    let rel_path = Path::new(rel);
-    if rel.is_empty()
-        || !rel_path
-            .components()
-            .all(|c| matches!(c, Component::Normal(_)))
-    {
-        anyhow::bail!("local action path escapes the repository");
-    }
-
-    let action_dir = repo_root.join(rel);
-    if workflow::open_child_dir_path(repo_root, rel_path)?.is_none() {
-        anyhow::bail!("{} is not a directory", action_dir.display());
-    }
-    Ok(action_dir)
-}
-
-fn scan_local_action_source(
-    repo_root: &Path,
-    action: &LocalActionRef,
-    collector: &mut AuditCollector,
-    config: &Config,
-) -> Result<ActionScanStatus> {
-    let action_dir = local_action_dir(repo_root, action)?;
-    let mut targets = collect_local_source_files(&action_dir)?;
-    force_include_local_action_entrypoints(&action_dir, &mut targets);
-    if targets.is_empty() {
-        return Ok(ActionScanStatus::Complete);
-    }
-
-    let mut complete = true;
-    for (path, kind) in targets {
-        let content = match std::fs::read_to_string(&path) {
-            Ok(content) => content,
-            Err(_) => {
-                complete = false;
-                continue;
-            }
-        };
-        let relative = path
-            .strip_prefix(&action_dir)
-            .unwrap_or(&path)
-            .to_string_lossy()
-            .replace('\\', "/");
-        let source_label = format!("{} ({relative})", action.path);
-        match kind {
-            SourceFileKind::ActionYml => match serde_norway::from_str::<Value>(&content) {
-                Ok(yaml) => {
-                    scan_action_yml_runs(&yaml, &source_label, &action.path, collector, config);
-                }
-                Err(_) => {
-                    complete = false;
-                }
-            },
-            SourceFileKind::JavaScript => {
-                scan_js_content(&content, &source_label, &action.path, collector, config);
-            }
-            SourceFileKind::Python => {
-                scan_py_content(&content, &source_label, &action.path, collector, config);
-            }
-            SourceFileKind::Dockerfile => {
-                scan_dockerfile_content(&content, &source_label, &action.path, collector, config);
-            }
-        }
-    }
-
-    Ok(ActionScanStatus::from_complete(complete))
-}
-
-async fn scan_action_source(
-    client: &GitHubClient,
-    action: &ActionRef,
-    collector: &mut AuditCollector,
-    config: &Config,
-) -> Result<ActionScanStatus> {
-    let action_name = format!("{}@{}", action.full_name(), short_sha(&action.ref_string));
-    let tree = client
-        .fetch_tree(&action.owner, &action.repo, &action.ref_string)
-        .await?;
-
-    let base = action.subpath.as_deref().unwrap_or("");
-    let mut targets = select_source_files(&tree, base);
-    if targets.is_empty() {
-        return Ok(ActionScanStatus::Complete);
-    }
-
-    // Fetch concurrently, then scan in tree order so findings are
-    // deterministic regardless of which fetch lands first. A failed fetch
-    // makes the scan incomplete, so the caller will not cache a clean verdict.
-    let (mut contents, mut complete) = fetch_remote_source_files(client, action, &targets).await;
-    let initial_len = targets.len();
-    force_include_remote_action_entrypoints(base, &mut targets, &contents);
-    if targets.len() > initial_len {
-        let (new_contents, new_complete) =
-            fetch_remote_source_files(client, action, &targets[initial_len..]).await;
-        contents.extend(new_contents);
-        complete &= new_complete;
-    }
-
-    for ((path, kind), content) in targets.iter().zip(contents) {
-        let content = match content {
-            Some(Ok(content)) => content,
-            _ => continue,
-        };
-        let source_label = format!("{} ({path})", action.full_name());
-        match kind {
-            SourceFileKind::ActionYml => match serde_norway::from_str::<Value>(&content) {
-                Ok(yaml) => {
-                    scan_action_yml_runs(&yaml, &source_label, &action_name, collector, config);
-                }
-                Err(_) => {
-                    complete = false;
-                }
-            },
-            SourceFileKind::JavaScript => {
-                scan_js_content(&content, &source_label, &action_name, collector, config);
-            }
-            SourceFileKind::Python => {
-                scan_py_content(&content, &source_label, &action_name, collector, config);
-            }
-            SourceFileKind::Dockerfile => {
-                scan_dockerfile_content(&content, &source_label, &action_name, collector, config);
-            }
-        }
-    }
-
-    Ok(ActionScanStatus::from_complete(complete))
-}
-
-fn scan_action_yml_runs(
-    yaml: &Value,
-    source_file: &str,
-    action_name: &str,
-    collector: &mut AuditCollector,
-    config: &Config,
-) {
-    // runs.steps[].run (composite actions), including nested parallel groups.
-    // base_line 0: positions are block-relative — the block is never located
-    // inside the fetched file.
-    if let Some(steps) = yaml.get("runs").and_then(|r| r.get("steps")) {
-        for run in collect_step_run_blocks(steps) {
-            scan_shell_content(run, source_file, 0, action_name, collector, config);
-        }
-    }
-
-    // runs.args (some actions use shell: bash with inline scripts)
-    if let Some(args) = yaml.get("runs").and_then(|r| r.get("args")) {
-        if let Some(args) = args.as_str() {
-            scan_shell_content(args, source_file, 0, action_name, collector, config);
-        } else if let Some(args) = args.as_sequence() {
-            for arg in args {
-                if let Some(arg) = arg.as_str() {
-                    scan_shell_content(arg, source_file, 0, action_name, collector, config);
-                }
-            }
-        }
-    }
-}
-
-fn short_sha(sha: &str) -> &str {
-    if sha.len() >= 7 { &sha[..7] } else { sha }
 }
 
 #[cfg(test)]
@@ -2305,58 +1250,6 @@ jobs:
 
         assert_eq!(c.findings.len(), 1);
         assert_eq!(c.findings[0].severity, "high");
-    }
-
-    #[test]
-    fn is_shell_comment_line_detects_leading_hash() {
-        assert!(is_shell_comment_line("# comment"));
-        assert!(is_shell_comment_line("    # indented comment"));
-        assert!(is_shell_comment_line("\t# tab indent"));
-    }
-
-    #[test]
-    fn is_shell_comment_line_rejects_trailing_hash() {
-        // Trailing comments aren't covered — an unquoted `#` can't be
-        // distinguished from one inside a string without real shell parsing.
-        assert!(!is_shell_comment_line("echo hello  # note"));
-        assert!(!is_shell_comment_line("foo=\"# not a comment\""));
-    }
-
-    #[test]
-    fn join_continuations_merges_trailing_backslash() {
-        let joined = join_continuations("curl https://x.example/s.sh \\\n  | sh\n");
-        assert_eq!(joined.len(), 1);
-        assert_eq!(joined[0].0, 0);
-        assert_eq!(joined[0].1, "curl https://x.example/s.sh  | sh");
-    }
-
-    #[test]
-    fn join_continuations_merges_trailing_pipeline_operators() {
-        let joined = join_continuations(
-            "curl https://x.example/s.sh |\n  sh\necho ok &&\n  true\necho no ||\n  false\n",
-        );
-        assert_eq!(
-            joined,
-            vec![
-                (0, "curl https://x.example/s.sh | sh".into()),
-                (2, "echo ok && true".into()),
-                (4, "echo no || false".into()),
-            ]
-        );
-    }
-
-    #[test]
-    fn join_continuations_leaves_unbroken_lines_alone() {
-        let joined = join_continuations("echo one\necho two\n");
-        assert_eq!(joined, vec![(0, "echo one".into()), (1, "echo two".into())]);
-    }
-
-    #[test]
-    fn join_continuations_handles_multiple_breaks() {
-        let joined = join_continuations("curl \\\n  -L \\\n  https://x.example/s.sh | sh\n");
-        assert_eq!(joined.len(), 1);
-        assert!(joined[0].1.contains("curl"));
-        assert!(joined[0].1.contains("| sh"));
     }
 
     #[test]
@@ -3201,412 +2094,6 @@ const d = require("node:https").get("https://example.com/install.sh", cb);
     }
 
     #[test]
-    fn scan_action_yml_composite_steps() {
-        let yaml: serde_norway::Value = serde_norway::from_str(
-            r#"
-runs:
-  using: composite
-  steps:
-    - run: curl -L https://example.com/install.sh -o install.sh
-"#,
-        )
-        .unwrap();
-        let mut c = AuditCollector::new(false);
-        scan_action_yml_runs(&yaml, "action.yml", "test-action", &mut c, &DEFAULT_CONFIG);
-        assert_eq!(c.findings.len(), 1);
-    }
-
-    #[test]
-    fn scan_action_yml_composite_parallel_steps() {
-        let yaml: serde_norway::Value = serde_norway::from_str(
-            r#"
-runs:
-  using: composite
-  steps:
-    - parallel:
-        - run: curl https://example.com/install.sh | sh
-"#,
-        )
-        .unwrap();
-        let mut c = AuditCollector::new(false);
-        scan_action_yml_runs(&yaml, "action.yml", "test-action", &mut c, &DEFAULT_CONFIG);
-        assert_eq!(c.findings.len(), 1);
-        assert_eq!(c.findings[0].severity, "high");
-    }
-
-    #[test]
-    fn scan_action_yml_args() {
-        let yaml: serde_norway::Value = serde_norway::from_str(
-            r#"
-runs:
-  using: node20
-  args: |
-    curl -L https://example.com/install.sh -o install.sh
-"#,
-        )
-        .unwrap();
-        let mut c = AuditCollector::new(false);
-        scan_action_yml_runs(&yaml, "action.yml", "test-action", &mut c, &DEFAULT_CONFIG);
-        assert_eq!(c.findings.len(), 1);
-    }
-
-    #[test]
-    fn scan_action_yml_sequence_args() {
-        let yaml: serde_norway::Value = serde_norway::from_str(
-            r#"
-runs:
-  using: node20
-  args:
-    - --flag
-    - curl -L https://example.com/install.sh -o install.sh
-"#,
-        )
-        .unwrap();
-        let mut c = AuditCollector::new(false);
-        scan_action_yml_runs(&yaml, "action.yml", "test-action", &mut c, &DEFAULT_CONFIG);
-        assert_eq!(c.findings.len(), 1);
-    }
-
-    #[test]
-    fn scan_action_yml_no_runs_key() {
-        let yaml: serde_norway::Value = serde_norway::from_str("name: test\n").unwrap();
-        let mut c = AuditCollector::new(false);
-        scan_action_yml_runs(&yaml, "action.yml", "test-action", &mut c, &DEFAULT_CONFIG);
-        assert!(c.findings.is_empty());
-    }
-
-    #[test]
-    fn short_sha_full() {
-        assert_eq!(
-            short_sha("abcdef1234567890abcdef1234567890abcdef12"),
-            "abcdef1"
-        );
-    }
-
-    #[test]
-    fn short_sha_short() {
-        assert_eq!(short_sha("abc"), "abc");
-    }
-
-    #[test]
-    fn remote_action_scan_key_includes_subpath() {
-        let mut first = ActionRef {
-            owner: "owner".to_string(),
-            repo: "repo".to_string(),
-            subpath: Some("a".to_string()),
-            ref_string: "abcdef1234567890abcdef1234567890abcdef12".to_string(),
-            ref_type: workflow::RefType::Sha,
-            tag_comment: None,
-            line_number: 1,
-            raw_line: String::new(),
-        };
-        let mut second = first.clone();
-        second.subpath = Some("b".to_string());
-
-        assert_ne!(
-            remote_action_scan_key(&first),
-            remote_action_scan_key(&second)
-        );
-        first.subpath = None;
-        assert_ne!(
-            remote_action_scan_key(&first),
-            remote_action_scan_key(&second)
-        );
-    }
-
-    // ── vendored-path filtering ────────────────────────────────────────
-
-    #[test]
-    fn vendored_path_skips_dependency_dirs() {
-        assert!(is_vendored_path("node_modules/lodash/index.js"));
-        assert!(is_vendored_path("dist/node_modules/x.js")); // nested
-        assert!(is_vendored_path(
-            ".venv/lib/python3.12/site-packages/requests/api.py"
-        ));
-        assert!(is_vendored_path("venv/bin/thing.py"));
-        assert!(is_vendored_path("tools/site-packages/pkg.py"));
-    }
-
-    #[test]
-    fn vendored_path_keeps_action_code() {
-        // The action's own bundled/source files must still be scanned.
-        assert!(!is_vendored_path("dist/index.js"));
-        assert!(!is_vendored_path("src/main.ts"));
-        assert!(!is_vendored_path("action.yml"));
-        assert!(!is_vendored_path("scripts/setup.py"));
-    }
-
-    #[test]
-    fn vendored_path_matches_whole_components_only() {
-        // Substrings and lookalike names must not be skipped.
-        assert!(!is_vendored_path("node_modules_helper.js"));
-        assert!(!is_vendored_path("my_vendor/index.js"));
-        assert!(!is_vendored_path("src/venvironment.py"));
-    }
-
-    // ── select_source_files ────────────────────────────────────────────
-
-    fn tree_entry(path: &str, entry_type: &str) -> crate::github::TreeEntry {
-        crate::github::TreeEntry {
-            path: path.into(),
-            entry_type: entry_type.into(),
-        }
-    }
-
-    #[test]
-    fn action_yml_entrypoint_paths_stay_in_action_subpath() {
-        let yaml: serde_norway::Value = serde_norway::from_str(
-            r#"
-runs:
-  using: node20
-  pre: ./preload
-  main: dist/runner
-  post: ../outside
-"#,
-        )
-        .unwrap();
-
-        assert_eq!(
-            action_yml_entrypoint_paths(&yaml, "actions/sub"),
-            vec![
-                "actions/sub/dist/runner".to_string(),
-                "actions/sub/preload".to_string()
-            ]
-        );
-        assert!(normalize_action_entrypoint_path("", "/tmp/runner").is_none());
-        assert!(normalize_action_entrypoint_path("", r"dist\\runner").is_none());
-    }
-
-    #[test]
-    fn select_source_files_classifies_and_filters_in_order() {
-        let tree = vec![
-            tree_entry("action.yml", "blob"),
-            tree_entry("dist/index.js", "blob"), // bundled action code — kept
-            tree_entry("dist/index.mjs", "blob"),
-            tree_entry("src/main.cjs", "blob"),
-            tree_entry("src/main.ts", "blob"),
-            tree_entry("setup.py", "blob"),
-            tree_entry("Dockerfile", "blob"),
-            tree_entry("README.md", "blob"), // not scannable
-            tree_entry("node_modules/dep/i.js", "blob"), // vendored — skipped
-            tree_entry("src", "tree"),       // directory entry — skipped
-        ];
-        let got = select_source_files(&tree, "");
-        assert_eq!(
-            got,
-            vec![
-                ("action.yml".to_string(), SourceFileKind::ActionYml),
-                ("dist/index.js".to_string(), SourceFileKind::JavaScript),
-                ("dist/index.mjs".to_string(), SourceFileKind::JavaScript),
-                ("src/main.cjs".to_string(), SourceFileKind::JavaScript),
-                ("src/main.ts".to_string(), SourceFileKind::JavaScript),
-                ("setup.py".to_string(), SourceFileKind::Python),
-                ("Dockerfile".to_string(), SourceFileKind::Dockerfile),
-            ]
-        );
-    }
-
-    #[test]
-    fn select_source_files_scopes_to_subpath_base() {
-        let tree = vec![
-            tree_entry("action-a/action.yml", "blob"),
-            tree_entry("action-a/index.js", "blob"),
-            tree_entry("action-b/action.yml", "blob"), // different subpath — excluded
-        ];
-        let got = select_source_files(&tree, "action-a");
-        assert_eq!(
-            got,
-            vec![
-                ("action-a/action.yml".to_string(), SourceFileKind::ActionYml),
-                ("action-a/index.js".to_string(), SourceFileKind::JavaScript),
-            ]
-        );
-    }
-
-    #[test]
-    fn select_source_files_subpath_base_requires_path_boundary() {
-        let tree = vec![
-            tree_entry("action-a/action.yml", "blob"),
-            tree_entry("action-a/index.js", "blob"),
-            tree_entry("action-abcd/action.yml", "blob"),
-            tree_entry("action-abcd/index.js", "blob"),
-        ];
-        let got = select_source_files(&tree, "action-a");
-        assert_eq!(
-            got,
-            vec![
-                ("action-a/action.yml".to_string(), SourceFileKind::ActionYml),
-                ("action-a/index.js".to_string(), SourceFileKind::JavaScript),
-            ]
-        );
-    }
-
-    #[test]
-    fn select_source_files_empty_when_nothing_scannable() {
-        let tree = vec![
-            tree_entry("README.md", "blob"),
-            tree_entry("LICENSE", "blob"),
-            tree_entry("node_modules/x/index.js", "blob"),
-        ];
-        assert!(select_source_files(&tree, "").is_empty());
-    }
-
-    #[test]
-    fn collect_local_source_files_filters_vendored_dirs() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let action_dir = dir.path().join(".github/actions/local");
-        std::fs::create_dir_all(action_dir.join("node_modules/dep")).unwrap();
-        std::fs::create_dir_all(action_dir.join("dist")).unwrap();
-        std::fs::create_dir_all(action_dir.join("src")).unwrap();
-        std::fs::write(action_dir.join("action.yml"), "name: local\n").unwrap();
-        std::fs::write(
-            action_dir.join("dist/helper.mjs"),
-            "fetch('https://example.com/x')",
-        )
-        .unwrap();
-        std::fs::write(
-            action_dir.join("dist/index.js"),
-            "fetch('https://example.com/x')",
-        )
-        .unwrap();
-        std::fs::write(
-            action_dir.join("src/main.cjs"),
-            "fetch('https://example.com/x')",
-        )
-        .unwrap();
-        std::fs::write(
-            action_dir.join("node_modules/dep/index.js"),
-            "fetch('https://evil.example/x')",
-        )
-        .unwrap();
-
-        let got = collect_local_source_files(&action_dir).unwrap();
-        let paths: Vec<_> = got
-            .iter()
-            .map(|(path, _)| {
-                path.strip_prefix(&action_dir)
-                    .unwrap()
-                    .to_string_lossy()
-                    .replace('\\', "/")
-            })
-            .collect();
-        assert_eq!(
-            paths,
-            vec![
-                "action.yml",
-                "dist/helper.mjs",
-                "dist/index.js",
-                "src/main.cjs"
-            ]
-        );
-    }
-
-    #[test]
-    fn scan_local_action_source_finds_composite_fetches() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let action_dir = dir.path().join(".github/actions/local");
-        std::fs::create_dir_all(&action_dir).unwrap();
-        std::fs::write(
-            action_dir.join("action.yml"),
-            r#"
-runs:
-  using: composite
-  steps:
-    - run: curl -fsSL https://example.com/install.sh | bash
-"#,
-        )
-        .unwrap();
-
-        let action = LocalActionRef {
-            path: "./.github/actions/local".to_string(),
-            line_number: 12,
-        };
-        let mut collector = AuditCollector::new(false);
-        let status =
-            scan_local_action_source(dir.path(), &action, &mut collector, &DEFAULT_CONFIG).unwrap();
-        assert_eq!(status, ActionScanStatus::Complete);
-        assert_eq!(collector.findings.len(), 1);
-        assert!(collector.findings[0].description.contains("piped to shell"));
-        assert_eq!(
-            collector.findings[0].source_file,
-            "./.github/actions/local (action.yml)"
-        );
-    }
-
-    #[test]
-    fn scan_local_action_source_includes_metadata_entrypoint() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let action_dir = dir.path().join(".github/actions/local");
-        std::fs::create_dir_all(action_dir.join("dist")).unwrap();
-        std::fs::write(
-            action_dir.join("action.yml"),
-            r#"
-runs:
-  using: node20
-  main: dist/runner
-"#,
-        )
-        .unwrap();
-        std::fs::write(
-            action_dir.join("dist/runner"),
-            r#"fetch("https://example.com/install")"#,
-        )
-        .unwrap();
-
-        let action = LocalActionRef {
-            path: "./.github/actions/local".to_string(),
-            line_number: 12,
-        };
-        let mut collector = AuditCollector::new(false);
-        let status =
-            scan_local_action_source(dir.path(), &action, &mut collector, &DEFAULT_CONFIG).unwrap();
-
-        assert_eq!(status, ActionScanStatus::Complete);
-        assert_eq!(collector.findings.len(), 1);
-        assert_eq!(
-            collector.findings[0].source_file,
-            "./.github/actions/local (dist/runner)"
-        );
-    }
-
-    #[test]
-    fn scan_local_action_source_rejects_parent_escape() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let action = LocalActionRef {
-            path: "./../outside".to_string(),
-            line_number: 1,
-        };
-        let mut collector = AuditCollector::new(false);
-        assert!(
-            scan_local_action_source(dir.path(), &action, &mut collector, &DEFAULT_CONFIG).is_err()
-        );
-    }
-
-    #[test]
-    fn scan_local_action_source_rejects_symlinked_action_root() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let outside = tempfile::TempDir::new().unwrap();
-        std::fs::write(outside.path().join("action.yml"), "name: outside\n").unwrap();
-        let actions_dir = dir.path().join(".github/actions");
-        std::fs::create_dir_all(&actions_dir).unwrap();
-        std::os::unix::fs::symlink(outside.path(), actions_dir.join("local")).unwrap();
-
-        let action = LocalActionRef {
-            path: "./.github/actions/local".to_string(),
-            line_number: 1,
-        };
-        let mut collector = AuditCollector::new(false);
-        let err = scan_local_action_source(dir.path(), &action, &mut collector, &DEFAULT_CONFIG)
-            .unwrap_err();
-
-        assert!(
-            err.to_string()
-                .contains("Refusing to scan symlinked directory")
-        );
-        assert!(collector.findings.is_empty());
-    }
-
-    #[test]
     fn py_scan_trusted_host_is_allowed() {
         let config = Config {
             trusted_hosts: vec!["api.example.com".to_string()],
@@ -3638,122 +2125,6 @@ runs:
         );
         assert_eq!(c.findings.len(), 1);
         assert_eq!(c.findings[0].action, "actions/checkout@abc1234");
-    }
-
-    #[tokio::test]
-    async fn scan_action_source_reports_incomplete_when_any_file_fetch_fails() {
-        use serde_json::json;
-        use wiremock::matchers::{method, path};
-        use wiremock::{Mock, MockServer, ResponseTemplate};
-
-        let server = MockServer::start().await;
-        Mock::given(method("GET"))
-            .and(path(
-                "/repos/o/r/git/trees/aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-            ))
-            .respond_with(ResponseTemplate::new(200).set_body_json(json!({
-                "tree": [
-                    { "path": "action.yml", "type": "blob" },
-                    { "path": "dist/index.js", "type": "blob" }
-                ]
-            })))
-            .mount(&server)
-            .await;
-        Mock::given(method("GET"))
-            .and(path("/repos/o/r/contents/action.yml"))
-            .respond_with(
-                ResponseTemplate::new(200)
-                    .set_body_string("runs:\n  using: node20\n  main: dist/index.js\n"),
-            )
-            .mount(&server)
-            .await;
-        Mock::given(method("GET"))
-            .and(path("/repos/o/r/contents/dist/index.js"))
-            .respond_with(ResponseTemplate::new(500))
-            .mount(&server)
-            .await;
-
-        let client = GitHubClient::with_base("t".into(), server.uri());
-        let action = ActionRef {
-            owner: "o".into(),
-            repo: "r".into(),
-            subpath: None,
-            ref_string: "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa".into(),
-            ref_type: workflow::RefType::Sha,
-            tag_comment: Some("v1.0.0".into()),
-            line_number: 1,
-            raw_line: String::new(),
-        };
-        let mut collector = AuditCollector::new(false);
-
-        let status = scan_action_source(&client, &action, &mut collector, &DEFAULT_CONFIG)
-            .await
-            .unwrap();
-
-        assert_eq!(status, ActionScanStatus::Incomplete);
-        assert!(
-            collector.findings.is_empty(),
-            "failed fetch should not invent findings, only block clean caching"
-        );
-    }
-
-    #[tokio::test]
-    async fn scan_action_source_includes_metadata_entrypoint() {
-        use serde_json::json;
-        use wiremock::matchers::{method, path};
-        use wiremock::{Mock, MockServer, ResponseTemplate};
-
-        let server = MockServer::start().await;
-        Mock::given(method("GET"))
-            .and(path(
-                "/repos/o/r/git/trees/aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-            ))
-            .respond_with(ResponseTemplate::new(200).set_body_json(json!({
-                "tree": [
-                    { "path": "action.yml", "type": "blob" },
-                    { "path": "dist/runner", "type": "blob" }
-                ]
-            })))
-            .mount(&server)
-            .await;
-        Mock::given(method("GET"))
-            .and(path("/repos/o/r/contents/action.yml"))
-            .respond_with(
-                ResponseTemplate::new(200)
-                    .set_body_string("runs:\n  using: node20\n  main: dist/runner\n"),
-            )
-            .up_to_n_times(1)
-            .mount(&server)
-            .await;
-        Mock::given(method("GET"))
-            .and(path("/repos/o/r/contents/dist/runner"))
-            .respond_with(
-                ResponseTemplate::new(200)
-                    .set_body_string(r#"fetch("https://example.com/install")"#),
-            )
-            .mount(&server)
-            .await;
-
-        let client = GitHubClient::with_base("t".into(), server.uri());
-        let action = ActionRef {
-            owner: "o".into(),
-            repo: "r".into(),
-            subpath: None,
-            ref_string: "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa".into(),
-            ref_type: workflow::RefType::Sha,
-            tag_comment: Some("v1.0.0".into()),
-            line_number: 1,
-            raw_line: String::new(),
-        };
-        let mut collector = AuditCollector::new(false);
-
-        let status = scan_action_source(&client, &action, &mut collector, &DEFAULT_CONFIG)
-            .await
-            .unwrap();
-
-        assert_eq!(status, ActionScanStatus::Complete);
-        assert_eq!(collector.findings.len(), 1);
-        assert_eq!(collector.findings[0].source_file, "o/r (dist/runner)");
     }
 
     #[test]
